fix errors in new system materials

+++ core/include/core/scene/geometry.h
@@ -11,24 +11,24 @@
 class Geometry : public Counter<Geometry>, Fixed {
 protected:
     Geometry() = delete;
-    Geometry(uint32_t vDeclId);
+    Geometry(uint16_t vDeclId);
     virtual ~Geometry();
 
 public:
-    uint32_t GetVDeclId() const noexcept { return m_vDeclId; }
+    uint16_t GetVDeclId() const noexcept { return m_vDeclId; }
 
     virtual void Bind(ContextPtr& context) = 0;
     virtual uint32_t Draw(ContextPtr& context, uint32_t firstInstanceIndex = 0) = 0;
 
 protected:
-    uint32_t m_vDeclId;
+    uint16_t m_vDeclId;
 };
 
 class VertexBuffer;
 class GeometryUnindexed final : public Geometry {
 public:
     GeometryUnindexed() = delete;
-    GeometryUnindexed(const std::shared_ptr<VertexBuffer>& vb, uint32_t vbOffsetBytes, uint32_t vbCount, uint32_t vDeclId);
+    GeometryUnindexed(const std::shared_ptr<VertexBuffer>& vb, uint32_t vbOffsetBytes, uint32_t vbCount, uint16_t vDeclId);
     ~GeometryUnindexed() final;
 
 public:
@@ -47,7 +47,7 @@ public:
     GeometryIndexed() = delete;
     GeometryIndexed(const std::shared_ptr<VertexBuffer>& vb, uint32_t vbOffsetBytes,
         const std::shared_ptr<IndexBuffer>& ib, uint32_t ibOffsetBytes, uint32_t ibCount, bool ibUint32,
-        uint32_t vDeclId);
+        uint16_t vDeclId);
     ~GeometryIndexed() final;
 
 public:
